Button fix and directory refactor
