+++ app/src/main/java/com/example/conductor/activities/MediaControllerInterfaceActivity.java
@@ -1,11 +1,10 @@
-package com.example.conductor;
+package com.example.conductor.activities;
 
 import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
 import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.Fragment;
 import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import android.Manifest;
@@ -36,6 +35,11 @@ import android.widget.ImageButton;
 import android.widget.SeekBar;
 import android.widget.TextView;
 
+import com.example.conductor.util.MusicController;
+import com.example.conductor.listeners.ProximityEventListener;
+import com.example.conductor.R;
+import com.example.conductor.listeners.SeekBarManager;
+import com.example.conductor.util.SpotifyHelper;
 import com.example.conductor.fragments.CameraFragment;
 import com.example.conductor.fragments.ShutterFragment;
