+++ app/src/main/java/com/example/conductor/listeners/ProximityEventListener.java
@@ -1,4 +1,4 @@
-package com.example.conductor;
+package com.example.conductor.listeners;
 
 import android.content.Context;
 import android.content.Intent;
