+++ app/src/main/java/com/example/conductor/listeners/SeekBarManager.java
@@ -1,13 +1,13 @@
-package com.example.conductor;
+package com.example.conductor.listeners;
 
-import android.content.ComponentName;
 import android.media.MediaMetadata;
 import android.media.session.MediaController;
-import android.media.session.MediaSessionManager;
-import android.os.Build;
 import android.util.Log;
 import android.widget.SeekBar;
 import android.widget.TextView;
+
+import com.example.conductor.R;
+import com.example.conductor.activities.MediaControllerInterfaceActivity;
 
 public class SeekBarManager implements SeekBar.OnSeekBarChangeListener {
