+++ app/src/main/java/com/example/conductor/util/GestureRecognition.java
@@ -1,4 +1,4 @@
-package com.example.conductor;
+package com.example.conductor.util;
 
 import android.content.Context;
