+++ app/src/main/java/com/example/conductor/util/MusicController.java
@@ -1,4 +1,4 @@
-package com.example.conductor;
+package com.example.conductor.util;
 
 import android.media.AudioManager;
