+++ app/src/main/res/layout/tutorial_page.xml
@@ -225,6 +225,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/back"
+        style="@style/ConductorButtonStyle"
+        android:layout_gravity="center"
         android:layout_marginTop="16dp"
         android:onClick="tutorialBackClicked" />
